requirements.txt
certifi==2022.9.24
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-WTF==1.0.1
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
nltk==3.7
numpy==1.23.5
pandas==1.5.1
pip==22.2.2
protobuf==3.20.1
python-dateutil==2.8.2
pytz==2022.6
regex==2022.10.31
scikit-learn==1.1.1
scipy==1.9.3
setuptools==65.5.0
six==1.16.0
threadpoolctl==3.1.0
tqdm==4.64.1
Werkzeug==2.2.2
wheel==0.37.1
wincertstore==0.2
WTForms==3.0.1
zipp==3.10.0
